Fix: Cast values to expected types

+++ test/Provider/nl_BE/PaymentTest.php
@@ -34,7 +34,7 @@ final class PaymentTest extends TestCase
         self::assertStringStartsWith('0', $numbers);
 
         // Mod97 check on first 8 digits
-        $checksum = 97 - fmod(substr($numbers, 0, 8), 97);
+        $checksum = 97 - fmod((float) substr($numbers, 0, 8), 97);
         // Make sure checksum is 2 characters long
         $checksum = sprintf('%02d', $checksum);
